chg: Added path Package
Made the "path" packaged and moved all directory and file functions
there.

BREAKING CHANGE

Packages there were in the root packed have been moved to the path
package, and some renamed.
