+++ .gitignore
@@ -8,3 +8,4 @@
 /build-version.json
 /git/testdata/repo-01/
 /git/testtmp/
+/path/testtmp/
